Convert Media Management settings to TypeScript
(cherry picked from commit radarr/radarr@2ebf391f85)
