+++ frontend/src/App/State/SettingsAppState.ts
@@ -18,10 +18,12 @@ import Notification from 'typings/Notification';
 import QualityProfile from 'typings/QualityProfile';
 import General from 'typings/Settings/General';
 import IndexerOptions from 'typings/Settings/IndexerOptions';
+import MediaManagement from 'typings/Settings/MediaManagement';
 import NamingConfig from 'typings/Settings/NamingConfig';
 import NamingExample from 'typings/Settings/NamingExample';
 import ReleaseProfile from 'typings/Settings/ReleaseProfile';
 import UiSettings from 'typings/Settings/UiSettings';
+import MetadataAppState from './MetadataAppState';
 
 type Presets<T> = T & {
   presets: T[];
@@ -36,6 +38,10 @@ export interface DownloadClientAppState
 
 export interface GeneralAppState
   extends AppSectionItemState<General>,
+    AppSectionSaveState {}
+
+export interface MediaManagementAppState
+  extends AppSectionItemState<MediaManagement>,
     AppSectionSaveState {}
 
 export interface NamingAppState
@@ -108,6 +114,8 @@ interface SettingsAppState {
   indexerOptions: IndexerOptionsAppState;
   indexers: IndexerAppState;
   languages: LanguageSettingsAppState;
+  mediaManagement: MediaManagementAppState;
+  metadata: MetadataAppState;
   naming: NamingAppState;
   namingExamples: NamingExamplesAppState;
   notifications: NotificationAppState;
