+++ frontend/src/typings/Settings/MediaManagement.ts
@@ -1,0 +1,23 @@
+export default interface MediaManagement {
+  autoUnmonitorPreviouslyDownloadedMovies: boolean;
+  recycleBin: string;
+  recycleBinCleanupDays: number;
+  downloadPropersAndRepacks: string;
+  createEmptyMovieFolders: boolean;
+  deleteEmptyFolders: boolean;
+  fileDate: string;
+  rescanAfterRefresh: string;
+  setPermissionsLinux: boolean;
+  chmodFolder: string;
+  chownGroup: string;
+  skipFreeSpaceCheckWhenImporting: boolean;
+  minimumFreeSpaceWhenImporting: number;
+  copyUsingHardlinks: boolean;
+  useScriptImport: boolean;
+  scriptImportPath: string;
+  importExtraFiles: boolean;
+  extraFileExtensions: string;
+  enableMediaInfo: boolean;
+  whisparrFolderLimit: number;
+  autoRenameFolders: boolean;
+}
